+++ siteupdate/cplusplus/classes/GraphGeneration/PlaceRadius.h
@@ -16,9 +16,9 @@ class PlaceRadius
 	std::string descr;	// long description of area, E.G. "New York City"
 	std::string title;	// filename title, short name for area, E.G. "nyc"
 	double lat, lng;	// center latitude, longitude
-	unsigned int r;		// radius in miles
+	double r;		// radius in miles
 
-	PlaceRadius(const char *, const char *, double &, double &, int &);
+	PlaceRadius(const char *, const char *, double &, double &, double &);
 
 	bool contains_vertex(HGVertex *);
 	bool contains_vertex(double, double);
